Welcome!
This is the GibHub repository for my personal website. I give credit to wiwet for designing most of the UI. The contact form was designed by myself. 

When time allows I'll create a captcha for the form to prevent potential bots from spamming. A personal blog is also in the pipeline for creation. 

Feel free to send me an email for any questions or comments.